💡refactor : MapApi 코드리팩토링
