+++ now-waypoint-core/src/main/java/isn_t_this_e_not_i/now_waypoint_core/domain/config/EmbeddedRedisConfig.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.data.redis.connection.RedisConfiguration;
 import redis.embedded.RedisServer;
 
 import java.io.IOException;
