+++ now-waypoint-core/src/test/java/isn_t_this_e_not_i/now_waypoint_core/domain/map/URLEncoderTest.java
@@ -1,4 +1,4 @@
-package isn_t_this_e_not_i.now_waypoint_core;
+package isn_t_this_e_not_i.now_waypoint_core.domain.map;
 
 import org.junit.jupiter.api.Test;
